requirements.txt
asgiref==3.6.0
Django==3.2.18
django-cors-headers==3.14.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.3.0
oauthlib==3.2.2
Pillow==9.5.0
pymongo==3.12.3
pytz==2023.3
requests_oauthlib==1.3.1
sqlparse==0.2.4
tzdata==2023.3
django-oauth-toolkit==1.5.0
